Convert serializers/mixin_yaml_test to rspec
